requirements/requirements.txt
allure-pytest==2.8.6
selenium==3.141.0
googletrans==2.4.0